move scopes from dockets to distribution_scopes (#20816)
* move scopes from dockets to distribution_scopes

* add comment, add rubocop disable
