+++ app/models/docket.rb
@@ -3,12 +3,13 @@
 class Docket
   include ActiveModel::Model
   include DistributionConcern
+  include DistributionScopes
 
   def docket_type
     fail Caseflow::Error::MustImplementInSubclass
   end
 
-  # rubocop:disable Metrics/CyclomaticComplexity
+  # rubocop:disable Metrics/CyclomaticComplexity, Metrics/PerceivedComplexity
   # :reek:LongParameterList
   def appeals(priority: nil, genpop: nil, ready: nil, judge: nil)
     fail "'ready for distribution' value cannot be false" if ready == false
@@ -27,7 +28,7 @@ class Docket
 
     scope.order("appeals.receipt_date")
   end
-  # rubocop:enable Metrics/CyclomaticComplexity
+  # rubocop:enable Metrics/CyclomaticComplexity, Metrics/PerceivedComplexity
 
   def count(priority: nil, ready: nil)
     # The underlying scopes here all use `group_by` statements, so calling
@@ -122,7 +123,7 @@ class Docket
   # rubocop:enable Metrics/MethodLength, Metrics/AbcSize, Lint/UnusedMethodArgument
 
   def self.nonpriority_decisions_per_year
-    Appeal.extending(Scopes).nonpriority
+    Appeal.extending(DistributionScopes).nonpriority
       .joins(:decision_documents)
       .where("decision_date > ?", 1.year.ago)
       .pluck(:id).size
@@ -148,89 +149,10 @@ class Docket
   end
 
   def docket_appeals
-    Appeal.where(docket_type: docket_type).extending(Scopes)
+    Appeal.where(docket_type: docket_type).extending(DistributionScopes)
   end
 
   def use_by_docket_date?
     FeatureToggle.enabled?(:acd_distribute_by_docket_date, user: RequestStore.store[:current_user])
-  end
-
-  module Scopes
-    include DistributionScopes
-
-    def priority
-      include_aod_motions
-        .where("advance_on_docket_motions.created_at > appeals.established_at")
-        .where("advance_on_docket_motions.granted = ?", true)
-        .or(include_aod_motions.where("people.date_of_birth <= ?", 75.years.ago))
-        .or(include_aod_motions.where("appeals.stream_type = ?", Constants.AMA_STREAM_TYPES.court_remand))
-        .group("appeals.id")
-    end
-
-    def nonpriority
-      include_aod_motions
-        .where("people.date_of_birth > ? or people.date_of_birth is null", 75.years.ago)
-        .where.not("appeals.stream_type = ?", Constants.AMA_STREAM_TYPES.court_remand)
-        .group("appeals.id")
-        .having("count(case when advance_on_docket_motions.granted "\
-          "\n and advance_on_docket_motions.created_at > appeals.established_at then 1 end) = ?", 0)
-    end
-
-    def include_aod_motions
-      joins(:claimants)
-        .joins("LEFT OUTER JOIN people on people.participant_id = claimants.participant_id")
-        .joins("LEFT OUTER JOIN advance_on_docket_motions on advance_on_docket_motions.person_id = people.id")
-    end
-
-    def ready_for_distribution
-      joins(:tasks)
-        .group("appeals.id")
-        .having("count(case when tasks.type = ? and tasks.status = ? then 1 end) >= ?",
-                DistributionTask.name, Constants.TASK_STATUSES.assigned, 1)
-    end
-
-    def genpop
-      joins(with_assigned_distribution_task_sql)
-        .with_original_appeal_and_judge_task
-        .where(
-          "appeals.stream_type != ? OR distribution_task.assigned_at <= ? OR original_judge_task.assigned_to_id in (?)",
-          Constants.AMA_STREAM_TYPES.court_remand,
-          CaseDistributionLever.cavc_affinity_days.days.ago,
-          JudgeTeam.judges_with_exclude_appeals_from_affinity
-        )
-    end
-
-    def with_original_appeal_and_judge_task
-      joins("LEFT JOIN cavc_remands ON cavc_remands.remand_appeal_id = appeals.id")
-        .joins("LEFT JOIN appeals AS original_cavc_appeal ON original_cavc_appeal.id = cavc_remands.source_appeal_id")
-        .joins(
-          "LEFT JOIN tasks AS original_judge_task ON original_judge_task.appeal_id = original_cavc_appeal.id
-           AND original_judge_task.type = 'JudgeDecisionReviewTask'
-           AND original_judge_task.status = 'completed'"
-        )
-    end
-
-    # Within the first 21 days, the appeal should be distributed only to the issuing judge.
-    def non_genpop_for_judge(judge)
-      joins(with_assigned_distribution_task_sql)
-        .with_original_appeal_and_judge_task
-        .where("distribution_task.assigned_at > ?", CaseDistributionLever.cavc_affinity_days.days.ago)
-        .where(original_judge_task: { assigned_to_id: judge.id })
-    end
-
-    def ordered_by_distribution_ready_date
-      joins(:tasks)
-        .group("appeals.id")
-        .order(
-          Arel.sql("max(case when tasks.type = 'DistributionTask' then tasks.assigned_at end)")
-        )
-    end
-
-    def non_ihp
-      joins(:tasks)
-        .group("appeals.id")
-        .having("count(case when tasks.type = ? then 1 end) = ?",
-                InformalHearingPresentationTask.name, 0)
-    end
   end
 end
